src/clean_and_label/requirements.txt
nltk
pandas
google-cloud-storage